requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-bcrypt==1.0.1
flask-babel==3.1.0
flask-talisman==1.1.0
flask-limiter==3.5.0
opencv-python==4.8.0.76
numpy==1.24.3
pillow==10.0.0
pycryptodome==3.19.0
python-dotenv==1.0.0
matplotlib==3.7.1
pandas==2.0.1
tabulate==0.9.0
